Moved channel modals to shared directory
