+++ contentcuration/contentcuration/frontend/channelList/router.js
@@ -1,11 +1,11 @@
 import VueRouter from 'vue-router';
 import ChannelList from './views/Channel/ChannelList';
-import ChannelModal from './views/Channel/ChannelModal';
-import ChannelDetailsModal from './views/Channel/ChannelDetailsModal';
 import ChannelSetList from './views/ChannelSet/ChannelSetList';
 import ChannelSetModal from './views/ChannelSet/ChannelSetModal';
 import CatalogList from './views/Channel/CatalogList';
 import { RouterNames, ListTypes } from './constants';
+import ChannelDetailsModal from 'shared/views/channel/ChannelDetailsModal';
+import ChannelModal from 'shared/views/channel/ChannelModal';
 
 const router = new VueRouter({
   routes: [
