+++ contentcuration/contentcuration/frontend/shared/views/channel/ChannelModal.vue
